[#809] Reduce code duplication in Xbase tests
Also fix package names in purexbase.tests and another NPE in the
XbaseCompiler
