+++ org.eclipse.xtext.purexbase.tests/META-INF/MANIFEST.MF
@@ -27,8 +27,6 @@ Import-Package: org.apache.commons.logging,
  org.junit.runners;version="4.12.0",
  org.junit.runners.model;version="4.12.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: org.eclipse.xtext.purexbase.test,
- org.eclipse.xtext.purexbase.test.data,
- org.eclipse.xtext.purexbase.tests;x-internal=true,
- org.eclipse.xtext.xbase.junit.evaluation
+Export-Package: org.eclipse.xtext.purexbase.tests;x-internal=true,
+ org.eclipse.xtext.purexbase.tests.data
 Automatic-Module-Name: org.eclipse.xtext.purexbase.tests
