--- org.eclipse.xtext.purexbase.tests/src/org/eclipse/xtext/purexbase/test/data/Shape.java
@@ -1,12 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012, 2016 itemis AG (http://www.itemis.eu) and others.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *******************************************************************************/
-package org.eclipse.xtext.purexbase.test.data;
-
-public interface Shape {
-}
