+++ org.eclipse.xtext.purexbase.tests/src/org/eclipse/xtext/purexbase/tests/data/Shape.java
@@ -1,27 +1,12 @@
 /*******************************************************************************
- * Copyright (c) 2013 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2012, 2016 itemis AG (http://www.itemis.eu) and others.
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.xtext.xbase.tests;
+package org.eclipse.xtext.purexbase.tests.data;
 
-import static org.junit.Assert.*;
-
-import org.junit.Ignore;
-import org.junit.Test;
-
-/**
- * @author dhuebner - Initial contribution and API
- */
-public class DummyTestClass {
-
-	@Test
-	@Ignore
-	public void test() {
-		fail("Not yet implemented");
-	}
-
+public interface Shape {
 }
